requirements.txt
pydantic
dynaconf
fire
moto[s3,sqs]
pytest
pytest-cov

pillow
torch
transformers

gradio
# monitor
sentry-sdk
